Integrate `forc-submit` into `forc-client`
