+++ Cargo.toml
@@ -7,7 +7,6 @@ members = [
     "forc-plugins/forc-doc",
     "forc-plugins/forc-fmt",
     "forc-plugins/forc-lsp",
-    "forc-plugins/forc-submit",
     "forc-plugins/forc-tx",
     "forc-test",
     "forc-tracing",
