+++ forc-plugins/forc-client/src/cmd/mod.rs
@@ -1,5 +1,7 @@
 pub mod deploy;
 pub mod run;
+pub mod submit;
 
 pub use deploy::Command as Deploy;
 pub use run::Command as Run;
+pub use submit::Command as Submit;
