+++ forc-plugins/forc-client/src/op/mod.rs
@@ -1,5 +1,7 @@
 mod deploy;
 mod run;
+mod submit;
 
 pub use deploy::deploy;
 pub use run::run;
+pub use submit::submit;
