+++ forc-plugins/forc-client/src/op/run.rs
@@ -19,8 +19,6 @@ use sway_core::BuildTarget;
 use tokio::time::timeout;
 use tracing::info;
 
-pub const NODE_URL: &str = "http://127.0.0.1:4000";
-
 pub struct RanScript {
     pub receipts: Vec<fuel_tx::Receipt>,
 }
@@ -66,7 +64,7 @@ pub async fn run_pkg(
         .node_url
         .as_deref()
         .or_else(|| manifest.network.as_ref().map(|nw| &nw.url[..]))
-        .unwrap_or(NODE_URL);
+        .unwrap_or(crate::default::NODE_URL);
     let client = FuelClient::new(node_url)?;
     let contract_ids = command
         .contract
