--- forc-plugins/forc-submit/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "forc-submit"
-version = "0.33.1"
-authors = ["Fuel Labs <contact@fuel.sh>"]
-edition = "2021"
-homepage = "https://fuel.network/"
-license = "Apache-2.0"
-repository = "https://github.com/FuelLabs/sway"
-description = "A `forc` plugin for submitting transactions to a Fuel network."
-
-[lib]
-path = "src/lib.rs"
-
-[[bin]]
-name = "forc-submit"
-path = "src/main.rs"
-
-[dependencies]
-anyhow = "1"
-chrono = { version = "0.4", default-features = false, features = ["std"] }
-clap = { version = "3", features = ["derive", "env"] }
-fuel-gql-client = { version = "0.15", default-features = false }
-serde = "1"
-serde_json = "1"
-tokio = { version = "1.8", features = ["macros", "rt-multi-thread"] }
